+++ src/configs/routeProps.ts
@@ -1,6 +1,7 @@
-import { RouteProps } from 'react-router';
-import { WatchList } from '~/src/pages/Watch/WatchList';
+import {RouteProps} from 'react-router';
+import {WatchList, WatchDetail} from '~/src/pages';
 
 export const routeProps: RouteProps[] = [
-    { path: '/', component: WatchList, exact: true }
+    {path: '/', component: WatchList, exact: true},
+    {path: '/detail/:id', children: WatchDetail}
 ];
