--- src/hooks/useApiClient.ts
@@ -1,6 +1,0 @@
-import { AxiosInstance } from 'axios';
-import { apiClient } from '~/src/services/service';
-
-export const useApiClient = (): AxiosInstance => {
-  return apiClient;
-}
