--- src/hooks/useAppContainer.ts
@@ -1,6 +1,0 @@
-import { useContext } from 'react'
-import { AppContext } from '~/src/contexts/AppContext'
-
-export const useAppContainer = () => {
-  return useContext(AppContext)
-}
