+++ src/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import { render } from 'react-dom';
-import {App} from './components/App';
+import {App} from './App';
 
 render(<App />, document.getElementById('root'));
