--- src/pages/Watch/WatchList/useWatchList.tsx
@@ -1,25 +1,0 @@
-import {useQuery} from '@bytesoftio/use-query'
-import {WatchListQueryParams} from '~/src/pages/Watch/WatchList/WatchListQueryParams'
-import {Watch} from '~/src/services/types/Watch';
-import {useAsync, AsyncHandle} from '@bytesoftio/use-async';
-import {SearchWatches} from '~/src/services/searchWatches';
-import {useApiClient} from '~/src/hooks/useApiClient';
-
-export const useWatchList = () => {
-    const [query, setQuery] = useQuery<WatchListQueryParams>({})
-    const payload = useWatches(query);
-    const handleSearch = (value: string) => {
-        setQuery({keyword: value})
-    };
-
-    return {
-        query,
-        payload,
-        handleSearch
-    };
-}
-const useWatches = (query: WatchListQueryParams): AsyncHandle<Watch[]> => {
-    const api = useApiClient();
-    return useAsync<Watch[]>(async () =>
-        (await SearchWatches(api, query.keyword) || []), [JSON.stringify(query)]);
-}
