+++ src/pages/WatchList/WatchList.scss
