+++ src/pages/index.tsx
@@ -1,0 +1,2 @@
+export {WatchList} from './WatchList/WatchList';
+export {WatchDetail} from './WatchDetail/WatchDetail';
