+++ src/shared/SearchBox/SearchBox.scss
