+++ src/shared/WatchCard/WatchCard.scss
