+++ src/shared/index.tsx
@@ -1,0 +1,2 @@
+export {SearchBox} from './SearchBox/SearchBox';
+export {WatchCard} from './WatchCard/WatchCard';
