+++ src/types/WatchListQueryParams.ts
